Switched HR page from callback to delayed dataget (no network dependency)

+++ client/lib/network/network.dart
@@ -213,12 +213,16 @@ class Network {
 
   // Tuple3 (good access, res status code, res message)
   static Future<Tuple3<bool, int, Map<String, dynamic>>> _serverGet(String route) async {
+    var timeout = OperationTimeoutTracker(const Duration(seconds: 10));
     Tuple3<bool, int, Map<String, dynamic>> response = const Tuple3(false, 0, {});
     var st = await getStates();
     if (st.item1 == NetworkHttpConnectionState.connected && st.item3 == SecurityState.secure) {
       final serverIp = await getServerIP();
       final uuid = await _http.getUuid();
       try {
+        if (timeout.isTimedOut) {
+          return const Tuple3(false, HttpStatus.requestTimeout, {});
+        }
         final serverRes = await http.get(Uri.parse('http://$serverIp:$requestPort/requests/$route/$uuid'));
         if (serverRes.body.isNotEmpty) {
           var decryptedM = await NetworkSecurity.decryptMessage(serverRes.body);
@@ -243,6 +247,7 @@ class Network {
 
   // Tuple3 (good access, res status code, res message in json)
   static Future<Tuple3<bool, int, Map<String, dynamic>>> _serverPost(String route, dynamic json) async {
+    var timeout = OperationTimeoutTracker(const Duration(seconds: 10));
     Tuple3<bool, int, Map<String, dynamic>> response = const Tuple3(false, 0, {});
     var st = await getStates();
     if (st.item1 == NetworkHttpConnectionState.connected && st.item3 == SecurityState.secure) {
@@ -250,6 +255,9 @@ class Network {
       final uuid = await _http.getUuid();
       try {
         var encryptedM = await NetworkSecurity.encryptMessage(json);
+        if (timeout.isTimedOut) {
+          return const Tuple3(false, HttpStatus.requestTimeout, {});
+        }
         final serverRes = await http.post(Uri.parse('http://$serverIp:$requestPort/requests/$route/$uuid'), body: encryptedM);
         if (serverRes.body.isNotEmpty) {
           var decryptedM = await NetworkSecurity.decryptMessage(serverRes.body);
@@ -274,6 +282,7 @@ class Network {
 
   // Tuple3 (good access, res status code, res message)
   static Future<Tuple3<bool, int, Map<String, dynamic>>> _serverDelete(String route, dynamic json) async {
+    var timeout = OperationTimeoutTracker(const Duration(seconds: 10));
     Tuple3<bool, int, Map<String, dynamic>> response = const Tuple3(false, 0, {});
     var st = await getStates();
     if (st.item1 == NetworkHttpConnectionState.connected && st.item3 == SecurityState.secure) {
@@ -281,6 +290,9 @@ class Network {
       final uuid = await _http.getUuid();
       try {
         var encryptedM = await NetworkSecurity.encryptMessage(json);
+        if (timeout.isTimedOut) {
+          return const Tuple3(false, HttpStatus.requestTimeout, {});
+        }
         final serverRes = await http.delete(Uri.parse('http://$serverIp:$requestPort/requests/$route/$uuid'), body: encryptedM);
         if (serverRes.body.isNotEmpty) {
           var decryptedM = await NetworkSecurity.decryptMessage(serverRes.body);
@@ -298,7 +310,6 @@ class Network {
 
   static Future<Tuple3<bool, int, Map<String, dynamic>>> serverDelete(String route, dynamic json) async {
     return await _serverDelete(route, json).timeout(const Duration(seconds: 15), onTimeout: () {
-      // Logger().e("Request Function Timeout: $route");
       return const Tuple3(false, HttpStatus.requestTimeout, {});
     });
   }
